trying to fix func
